주소 입력을 통해 수정페이지에 접근하는 경우 차단
작성자와 다른 로그인아이디인 경우 수정 및 삭제 버튼을 제거했습니다.

그러나 직접 접근을 하는 경우에 대해서도 작성자와 현재사용자가 같은지 확인하는 로직을 추가했습니다.

+++ src/main/java/codesquad/springcafe/article/ArticleController.java
@@ -55,23 +55,39 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{articleId}/form")
-    public String showEditArticleForm(@PathVariable long articleId, Model model) {
+    public String showEditArticleForm(@PathVariable long articleId, Model model, HttpServletRequest request) {
         Article article = articleDatabase.getArticle(articleId);
+        if (!isWriter(article.getWriter(), request)) {
+            return "redirect:/error/errorPage";
+        }
         model.addAttribute("article", article);
         return "article/editForm";
     }
 
     @PutMapping("/articles/{articleId}")
-    public String editArticle(@ModelAttribute Article editedArticle, @PathVariable long articleId) {
+    public String editArticle(@ModelAttribute Article editedArticle, @PathVariable long articleId, HttpServletRequest request) {
         editedArticle.setArticleId(articleId);
-        articleDatabase.updateArticle(editedArticle);
+        Article article = articleDatabase.getArticle(articleId);
+        if (!isWriter(article.getWriter(), request)) {
+            return "redirect:/error/errorPage";
+        }
 
+        articleDatabase.updateArticle(editedArticle);
         return "redirect:/articles/" + articleId;
     }
 
     @DeleteMapping("/articles/{articleId}")
-    public String deleteArticle(@PathVariable long articleId) {
+    public String deleteArticle(@PathVariable long articleId, HttpServletRequest request) {
+        Article article = articleDatabase.getArticle(articleId);
+        if (!isWriter(article.getWriter(), request)) {
+            return "redirect:/error/errorPage";
+        }
         articleDatabase.deleteArticle(articleId);
         return "redirect:/";
+    }
+
+    private boolean isWriter(String writer, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        return writer.equals(session.getAttribute("loginUserId"));
     }
 }
